Improve `Text` package unit tests and coverage
