+++ tests/Text/fixtures/inline.html
@@ -1,0 +1,6 @@
+<em>italic</em><br />
+<br />
+<strong>bold</strong><br />
+<br />
+This is some text<br />
+and some more text 👀.
