+++ tests/Text/fixtures/inline.md
@@ -1,0 +1,6 @@
+*italic*
+
+**bold**
+
+This is some text
+and some more text 👀.
